+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PMPDPlayGround">
-        <activity
-            android:name="com.tijanidian.ex06form.presentation.Ut02ex06Activity"
-            android:exported="true" />
+
         <activity
             android:name="com.tijanidian.mvvm.presentation.MvvmActivity"
             android:exported="true" />
@@ -23,7 +21,10 @@
             android:exported="true" />
         <activity
             android:name="ut2.UiTestGlovoActivity"
-            android:exported="true">
+            android:exported="true"/>
+            <activity
+                android:name="com.tijanidian.ut02ex06form.presentation.Ut02ex06Activity"
+                android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
