+++ app/src/main/java/com/tijanidian/ut02ex06form/domain/Models.kt
@@ -1,0 +1,5 @@
+package com.tijanidian.ut02ex06form.domain
+
+
+class PlayerModel (id:String,name:String) {
+}
