+++ app/src/main/java/com/tijanidian/ut02ex06form/presentation/FormFragment.kt
@@ -1,11 +1,10 @@
-package com.tijanidian.ex06form.presentation
+package com.tijanidian.ut02ex06form.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.tijanidian.pmpd_playground.R
 import com.tijanidian.pmpd_playground.databinding.ActivityFormFragmentBinding
 
 class FormFragment : Fragment() {
@@ -18,6 +17,10 @@ class FormFragment : Fragment() {
     ): View? {
         binding = ActivityFormFragmentBinding.inflate(inflater, container, false)
         return binding.root
+    }
+
+    fun save(){
+        binding.inputName.text
     }
 
     companion object {
