+++ app/src/main/java/com/tijanidian/ut02ex06form/presentation/FromFragmentViewModel.kt
@@ -1,0 +1,4 @@
+package com.tijanidian.ut02ex06form.presentation
+
+class FromFragmentViewModel {
+}
