+++ app/src/main/res/layout/view_card_glovo.xml
@@ -9,7 +9,7 @@
         android:layout_height="match_parent"
         android:layout_centerHorizontal="true"
         android:scaleType="centerCrop"
-        android:src="@mipmap/glovo_comida" />
+        android:src="@drawable/glovo_comida" />
 
     <TextView
         android:id="@+id/main_brand_text"
